# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for EduCRM: JWT auth (access + refresh), users, teachers, rooms, groups, schedules, attendance, grades, payments, files, notifications, dashboard, and AI analytics. Success responses use a JSON envelope: `{\"success\":true,\"data\":...}`. Errors use `{\"success\":false,\"error\":{\"code\",\"message\",\"kind\"}}`.",
        "title": "EduCRM API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@educrm.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/ai/analytics/admin-daily-summary": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-analytics"
                ],
                "summary": "AI admin daily summary",
                "parameters": [
                    {
                        "description": "Optional as_of RFC3339",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/analytics/debtors-summary": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-analytics"
                ],
                "summary": "AI debtors summary",
                "parameters": [
                    {
                        "description": "Optional month YYYY-MM",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/analytics/low-attendance": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-analytics"
                ],
                "summary": "AI low attendance summary",
                "parameters": [
                    {
                        "description": "Optional from/to RFC3339",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/analytics/student-warnings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Student uses self; staff must pass student_id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-analytics"
                ],
                "summary": "AI student warning suggestions",
                "parameters": [
                    {
                        "description": "student_id for staff",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/ai/analytics/teacher-recommendations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Teacher uses linked profile; staff must pass teacher_id (teachers table id).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-analytics"
                ],
                "summary": "AI teacher recommendations",
                "parameters": [
                    {
                        "description": "teacher_id for staff",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsFilters": {
            "type": "object",
            "properties": {
                "as_of": {
                    "description": "RFC3339 — admin daily, optional debtors anchor",
                    "type": "string",
                    "example": "2025-04-08T09:00:00Z"
                },
                "from": {
                    "description": "RFC3339 — attendance window start",
                    "type": "string",
                    "example": "2025-03-01T00:00:00Z"
                },
                "month": {
                    "description": "YYYY-MM — debtors billed month (default: current UTC month)",
                    "type": "string",
                    "example": "2025-04"
                },
                "student_id": {
                    "description": "staff: student user id for warnings",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "teacher_id": {
                    "description": "staff: teachers.id for recommendations",
                    "type": "string",
                    "example": "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                },
                "to": {
                    "description": "RFC3339 — attendance window end",
                    "type": "string",
                    "example": "2025-04-01T00:00:00Z"
                }
            }
        },
        "github_com_educrm_educrm-backend_internal_delivery_http_dto.AIAnalyticsResponse": {
            "type": "object",
            "properties": {
                "output": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "github_com_educrm_educrm-backend_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_educrm_educrm-backend_pkg_response.ErrorBody"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_educrm_educrm-backend_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}